Move getSiteBlocksLastPage to state/reader
