+++ client/me/site-blocks/main.jsx
@@ -19,7 +19,7 @@ import QuerySiteBlocks from 'components/data/query-site-blocks';
 import getBlockedSites from 'state/reader/site-blocks/selectors/get-blocked-sites';
 import isFetchingSiteBlocks from 'state/selectors/is-fetching-site-blocks';
 import getSiteBlocksCurrentPage from 'state/reader/site-blocks/selectors/get-site-blocks-current-page';
-import getSiteBlocksLastPage from 'state/selectors/get-site-blocks-last-page';
+import getSiteBlocksLastPage from 'state/reader/site-blocks/selectors/get-site-blocks-last-page';
 import SiteBlockListItem from './list-item';
 import InfiniteList from 'components/infinite-list';
 import { requestSiteBlocks } from 'state/reader/site-blocks/actions';
