+++ client/state/selectors/test/get-site-blocks-last-page.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import getSiteBlocksLastPage from 'state/selectors/get-site-blocks-last-page';
+import getSiteBlocksLastPage from 'state/reader/site-blocks/selectors/get-site-blocks-last-page';
 
 describe( 'getSiteBlocksLastPage()', () => {
 	test( 'should return the last page', () => {
